quimb_bench: merge algorithm scripts into their test files
Each script's run_one(chi, L) (or run_one_dense/run_one_symmetric/
run_one_jax/run_one_torch, where a file holds more than one variant)
now lives directly in its test_<name>.py file instead of a separate,
non-test module imported via a sys.path.insert(0,
os.path.join(os.path.dirname(__file__), "..")) hack. test_dmrg.py keeps
the dense and U(1)-symmetric DMRG cases together in one file, since
quimb's symmetric variant runs imaginary-time evolution from an
unconverged random seeded state to exercise block-sparse contraction
cost rather than to reach a ground-state energy comparable across
libraries.

Every benchmark test now asserts the returned energy against a
REFERENCE_ENERGIES dict (or DENSE_REFERENCE_ENERGIES/
SYMMETRIC_REFERENCE_ENERGIES/JAX_REFERENCE_ENERGIES/
TORCH_REFERENCE_ENERGIES) covering the full (chi, L) grid, replacing
the previous single fixed-point regression check plus an
isfinite-only full-grid sweep. The grid is parametrized with two
stacked @pytest.mark.parametrize("chi", ...) /
@pytest.mark.parametrize("length", ...) decorators instead of a single
parametrize over a combined (chi, length) list, and limit_memory is
applied as a plain decorator on a dedicated, parameter-free *_memory
test pinned to the (chi=16, L=20) point instead of being attached via
pytest.param(..., marks=...).
